+++ src/client/main.js
@@ -1,0 +1,10 @@
+import PlayerName from './PlayerName.svelte';
+
+const playerName = new PlayerName({
+	target: document.querySelector('section#player-name'),
+	props: {
+		socket,
+	}
+});
+
+export default playerName;