+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - begin life cycle for round
 - deck functionality
 - add/remove/update user logic
+- connect svelte
+- playerName component functionality
 
 ## [v0.5.0] - 2020-06-03
 ### Added